Finishing up some last syscall modifications that allow for printing of strings, right now there is no clearly defined end point (prints all strings in data section at once). Big endian storage is implemented, but may need to be changed

+++ src/MemoryLoader.java
@@ -17,7 +17,7 @@ public class MemoryLoader {
                 }
 
                 // Parse the instruction
-                int instruction = Integer.parseInt(line.replaceFirst("0x", ""), 16);
+                int instruction = (int) Long.parseLong(line.replaceFirst("0x", ""), 16);
 
                 // Store into text memory
                 state.textMemory.put(address, instruction);
